==> PythonScripts/1_DescriptiveAnalysis.ipynb <==
Importing Libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

df = pd.read_csv(r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket\02_Data\OriginalData\orders.csv', index_col = False)

path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

path

df = pd.read_csv(os.path.join(path, '02_data', 'OriginalData', 'orders.csv'), index_col = False)

df.head()

df.tail()

df.shape

df.info()

df.dtypes

df.columns

df.describe()

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02_data', 'OriginalData', 'orders.csv'), usecols = vars_list)

df.head()
