==> PythonScripts/2_DataWrangling.ipynb <==
# 2. Data Wrangling

#### Import libraries and dataframe
#### Dropping Columns
#### Renaming Columns
#### Transposing Dataframes
#### Creating Data Dictionaries
#### Creating Subset Data
#### Performing Value Counts

Importing Libraries and Dataframes

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

df_orders = pd.read_csv(os.path.join(path,'02_Data', 'OriginalData', 'orders.csv'), index_col = False) 

df_products = pd.read_csv(os.path.join(path,'02_Data', 'OriginalData', 'products.csv'), index_col = False) 

Dropping unneccessary columns from dataframe:

df_orders = df_orders.drop(columns = ['eval_set'])

df_orders.head()

Renaming column 'order_dow' to something more descriptive: 

df_orders.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

Transposing the 'departments' dataframe:

# Importing 'departments.csv' data
df_dptmt = pd.read_csv(os.path.join(path,'02_Data', 'OriginalData', 'departments.csv'), index_col = False)

# Transpose the wide format to long format.
df_dptmt.head()

df_dptmt.T

# Overwrite the dataframe itself.
df_dptmt_t = df_dptmt.T

df_dptmt_t

df_dptmt_t.reset_index()

Removing the first row and replacing with the proper header:

# Take the first row of df_dptmt_t for the header
new_header = df_dptmt_t.iloc[0]

print(new_header)

df_dptmt_t_new = df_dptmt_t[1:]

df_dptmt_t_new

df_dptmt_t_new.columns = new_header

df_dptmt_t_new

Creating a data dictionary for deparment ID and department names: 

data_dict = df_dptmt_t_new.to_dict('index')

data_dict

df_products.head()

# To look up what department 19 is, use data dictionary.
print(data_dict.get('19'))

Creating a subset showing products only belonging to department ID 19:

# Creating a subset of product names only belonging to department_id 19.
df_snacks = df_products[df_products['department_id'] == 19]

df_snacks

Performing a value count of total orders made for each hour of day:  

# Using count function to see frequency of orders made every order time. 
df_orders['order_time_of_day'].value_counts(dropna = False)

The busiest hour of the day for orders is 10am. In general, the busiest time frame for orders is 10am-3pm. 

Creating a subset of only breakfast items by showing only products belonging to department ID 14: 

# Creating a subset of products only belonging to the 'breakfast' department. 
df_products_breakfast = df_products[df_products['department_id'] == 14]

df_products_breakfast.head()

Creating a subset of items related to dinner parties by showing products from alcohol, deli, beverages, and meat/seafood departments:

# Creating a subset of products belonging only to relevant deparments. 
df_dinnerparty = df_products[df_products['department_id'].isin([5,20,7,12])]

# Performing value_count() to see if the only needed department_id's are in this new dataframe.  
df_dinnerparty['department_id'].value_counts(dropna = False)

# Confirming the included department_id numbers are in fact for the department names I need for this subset dataframe. 

print(data_dict.get('7'))
print(data_dict.get('20'))
print(data_dict.get('5'))
print(data_dict.get('12'))

# Confirming all other columns are included in each observations in this new subset data frame. 
df_dinnerparty.head()

Exporting df_orders dataframe as "orders_wrangled.csv" in PreparedData folder.

df_orders.to_csv(os.path.join(path, '02_Data', 'PreparedData', 'orders_wrangled.csv'))

Exporting df_dptmt_t_new dataframe as "departments_wrangled.csv" in PreparedData folder. 

df_dptmt_t_new.to_csv(os.path.join(path, '02_Data', 'PreparedData', 'departments_wrangled.csv'))
