==> PythonScripts/3_Missing&DuplicateValues.ipynb <==
# 3. Addressing Missing & Duplicate Values

Importing Libraries & Data

import pandas as pd
import numpy as np
import os

# Creating file path to get to Instacart Project file. 
path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

# Importing products.csv dataframe from "OriginalData" folder.
# Importing wrangled data orders.csv from "PreparedData" folder.
df_products = pd.read_csv(os.path.join(path,'02_Data', 'OriginalData', 'products.csv'), index_col = False) 
df_orders = pd.read_csv(os.path.join(path,'02_Data', 'PreparedData', 'orders_wrangled.csv'), index_col = False) 

Finding and addressing missing values:

# To find missing values in dataframe.
df_products.isnull().sum()

# To see only the null values, create a subset meeting condition of isnull() = True, and call for the new subset df. 
df_nan = df_products[df_products['product_name'].isnull() == True]

df_nan

# Check the size of df before removing missing values.
df_products.shape

df_products_clean = df_products[df_products['product_name'].isnull() == False]

# Check size of new cleaned df to make sure exactly 16 rows are removed. 
df_products_clean.shape

Another way to drop missing values:

# This will overwrite the original dataframe and replace with new df with missing value rows dropped. 
df_products.dropna(subset = ['product_name'], inplace = True)

Finding and addressing duplicates:

# This will create a new subset containing of only rows that are duplicated.
df_dups = df_products_clean[df_products_clean.duplicated()]

df_dups

Creating a new dataframe without the duplicates: 

# Check number of rows in df (this case it is that df_products_clean).
df_products_clean.shape

# Then, create a new dataframe that doesn’t include the duplicates identified above.
df_products_clean_no_dups = df_products_clean.drop_duplicates()

# Check the number of rows in the new df, confirm that only x number of rows identified earlier are removed, which should be
# 5 less than the original dataframe in this case. 
df_products_clean_no_dups.shape

Exporting to CSV after Addressing Missing & Duplicate Values

df_products_clean_no_dups.to_csv(os.path.join(path,'02_Data', 'PreparedData', 'products_checked.csv'))

Running descriptive analysis on the 'products' dataframe:

df_products.describe()

df_products_clean_no_dups.describe()

Things to take note: in product_id, there are 49672 enteries (or 49677 before dropping duplicates), but the max value is 49688, which means is that when we dropped the 16 rows with missing values for product_name, it created holes in the consecutive numbering of the ids. This is not really a problem, but important to note here. Other issue is that the max value for prices is $99,9999, this value is far greater than most other pricing, plus suspicious since we are dealing with groceries item. This price could be a typo, we need to look further into this. 

Checking for mixed-type data in the 'orders' dataframe:

# This for-loop finds mixed-type data in the dataframe.
for col in df_orders.columns.tolist():
    weird = (df_orders[[col]].applymap(type) != df_orders[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_orders[weird]) > 0:
        print(col)

Since there was no output, there was not a column with mixed-type data.

Running data type test to see what each column has as data type.:

df_orders.dtypes

Checking for missing values in the 'orders' dataframe: 

# Checking df_orders dataframe for missing observations. 
df_orders.isnull().sum()

We have 206,209 missing values (of total 3,421,083 observations) in 'days_since_prior_order' column, which isn't odd since the first order for each new user will have 'NaN' recorded in this column. If anything, this number should match how many users (or total count of "user_id") there are. 

df_orders.shape

# Running basic stats to see what max value user_id is.
df_orders.describe()

Sure enough, the largest number for user_id is 206209; thus it makes sense that there are 206,209 "NaN" enteries under the 'days_since_prior_order' column. Since this is valuable information, the "NaN" enteries will not be removed or changed. 

Checking for duplicates in 'orders' dataframe: 

# Finding dupicates in df_orders dataframe and creating a subset of them.
df_orders_dups = df_orders[df_orders.duplicated()]

df_orders_dups

Since the new subset meant for duplicates has no rows, there were no duplicates. 

Exporting dataframe to csv:

df_orders.to_csv(os.path.join(path,'02_Data', 'PreparedData', 'orders_checked.csv'))
