==> PythonScripts/4_MergingDataI.ipynb <==
# 4. Merging Data (Part I)

Importing libraries:

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

Importing the first dataframe: 

df_ords_prior = pd.read_csv(r"C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket\02_Data\OriginalData\order_products__prior.csv", index_col = False) 

df_ords_prior.head()

Importing the second dataframe: 

df_ords = pd.read_csv(os.path.join(path,'02_Data', 'PreparedData', 'orders_checked.csv'), index_col = False) 

df_ords.head()

df_ords_prior.shape

df_ords.shape

Dropping unneccessary columns from dataframes:

df_ords = df_ords.drop(columns = ['Unnamed: 0'])

df_ords = df_ords.drop(columns = ['Unnamed: 0.1'])

df_ords.head()

df_ords.shape

Merging the two dataframes together by inner join: 

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', how = 'inner', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

Exporting the newly merged dataframe to pickle file: 

df_merged_large.to_pickle(os.path.join(path, '02_Data','PreparedData', 'orders_products_combined.pkl'))
