==> PythonScripts/5_MergingDataII.ipynb <==
# 5. Merging Data (Part II)

Importing libraries:

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

Importing the 'order_products_combined' dataframe from the pickle file:

orders_products_combined = pd.read_pickle(os.path.join(path,'02_Data', 'PreparedData', 'orders_products_combined.pkl')) 

orders_products_combined.shape

orders_products_combined.head()

The shape is confirmed to be exactly as how it was exported.

Importing the 'products' dataframe to combine with the 'orders_products_combined' dataframe:  

# Import products data set from "PreparedData" folder.
df_products = pd.read_csv(os.path.join(path,'02_Data', 'PreparedData', 'products_checked.csv'), index_col = False) 

df_products.head()

# Dropping unneeded column 'Unnamed: 0'
df_products = df_products.drop(columns = ['Unnamed: 0'])

# Checking to see if successful.
df_products.head()

df_products.shape

# Before merging the two df, dropping '_merge' column from orders_products_combined dataframe since it isn't needed. 
orders_products_combined = orders_products_combined.drop(columns = ['_merge'])

orders_products_combined.head()

Merging the 'products' dataframe to the 'orders_products_combined' dataframe by inner joining on the variable 'product_id' as the key: 

df_merged_ords_productName = orders_products_combined.merge(df_products, on = 'product_id', how = 'inner', indicator = True)

Confirming the results of the merge a value count on the merge flag: 

# Using count of values in '_merge' column to see if full merge. 
df_merged_ords_productName['_merge'].value_counts()

Since counts are made only for 'both' enteries, the combining was a full merge. 

Exporting the newly merged dataframe as orders_products_merged in pickle format: 

df_merged_ords_productName.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'orders_products_merged.pkl'))
