==> PythonScripts/6_CreatingVariables.ipynb <==
# 6. Creating New Variables

#### Importing libraries and dataframe
#### Creating a new column named 'price_range_loc'
#### Creating a new column named 'busiest_day'
#### Creating a new column named "busiest_period_of_day"

Importing Libraries and Dataframe:

# Import libraries
import pandas as pd
import numpy as np
import os

# Create path
path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

# Import dataframe saved as pickel
ords_prods_merged = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'orders_products_merged.pkl'))

# Create a subset to avoid potential issues dealing with user-defined functions
df = ords_prods_merged[:1000000]

df.shape

df.head()

Creating a new column named 'price_range_loc' by using loc() function on whole dataframe:

ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High Range Product'

ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid Range Product'

ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low Range Product'

ords_prods_merged['price_range_loc'].value_counts(dropna = False)

Creating new column named 'busiest_day':

# Goal is to create a new column that summarizes how busy each day of the week is.
# First, we find out which day has the most orders by using value_count
ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Creating a user-defined function to create this new column: 

result = []

for value in ords_prods_merged['orders_day_of_week']:
    
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regular busy')

# Combine the new array named 'result' onto the dataframe
ords_prods_merged['busiest_day'] = result

# Check value count in the newly created column "busiest_day"
ords_prods_merged['busiest_day'].value_counts(dropna = False)

Creating a new column named "busiest_period_of_day" identify busiest hours of the day, hours will be labeled by periods of time, as "Most orders," "Average orders," or "Fewest orders" :

# Identify busiest hours
ords_prods_merged['order_time_of_day'].value_counts(dropna = False)

We need to categorize the hours into three time-frames: 
"Most orders" time-frame is 9-16 (9am - 4pm) with >2.4 million order per hour
"Average orders" time-frame is 7-8 (7am - 8am) & 17-23 (5pm - 11pm) with orders between 300K ~ 2400K
"Fewest orders" time-frame is 0-6 (12am - 6am) with <300K orders per hour

# loc() function to categorize 0-6 time-frame:
ords_prods_merged.loc[ords_prods_merged['order_time_of_day'] <= 6, 'busiest_period_of_day'] = 'fewest orders'

# loc() function to categorize 7-8 time-frame:
ords_prods_merged.loc[(ords_prods_merged['order_time_of_day'] > 6) & (ords_prods_merged['order_time_of_day'] < 9), 'busiest_period_of_day'] = 'average orders'

# loc() function to categorize 17-24 time-frame:
ords_prods_merged.loc[(ords_prods_merged['order_time_of_day'] > 16) & (ords_prods_merged['order_time_of_day'] < 24), 'busiest_period_of_day'] = 'average orders'

# loc() function to categorize 9-16 time-frame:
ords_prods_merged.loc[(ords_prods_merged['order_time_of_day'] > 8) & (ords_prods_merged['order_time_of_day'] < 17), 'busiest_period_of_day'] = 'most orders'

Checking the value count of this new column:

ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

# Checking head() to see what the column looks like. 
ords_prods_merged.head()

Exporting the dataframe as a pickle file into the "PreparedData" folder:

ords_prods_merged.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'orders_products_merged.pkl'))
