==> PythonScripts/9_Crosstabs.ipynb <==
# 9. Creating Crosstabs

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'orders_products_merged.pkl'))

crosstab = pd.crosstab(ords_prods_merge['days_since_prior_order'], ords_prods_merge['order_number'], dropna = False)

crosstab.to_clipboard()
