==> PythonScripts/10.1_FinalAnalysis.ipynb <==
# 10.1 FinalAnalysis

Import the data set exported last script.

import pandas as pd
import numpy as np
import os

path = path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

ords_prods_customers = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'ords_prods_customers.pkl'))

ords_prods_customers.info()

Considering security implications that might exist for most recent data. The only columns with PII data are 'first_name' and 'last_name'. Since customer names are not needed, creating a new dataframe with the two columns dropped. If we do, in some rare case, need to go back to trace the names of the customers, we will still have customer_id's to do so. 

df_names_dropped = ords_prods_customers.drop(columns = ['first_name', 'last_name'])

df_names_dropped.info()

To investigate into customer behavior in different geographic areas, creating a regional segmentation of the data and categorized into 'region' in new column based on the 'state' column from your customers data set. 

# Creating lists for each region
Northeast_states = ['Maine', 'New Hamphire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 
                    'New York', 'Pennsylvania', 'New Jersey']

Midwest_states = ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 
                  'Minnesota', 'Iowa', 'Missouri']

South_states = ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina',
                'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 
                'Louisiana']

West_states = ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington',
               'Oregon', 'California', 'Hawaii']

# Assign regional names as enteries in new column named 'region'
df_names_dropped.loc[df_names_dropped['state'].isin(West_states), 'region'] = 'West'

df_names_dropped.loc[df_names_dropped['state'].isin(South_states), 'region'] = 'South'

df_names_dropped.loc[df_names_dropped['state'].isin(Midwest_states), 'region'] = 'Midwest'

df_names_dropped.loc[df_names_dropped['state'].isin(Northeast_states), 'region'] = 'Northeast'

df_names_dropped['region'].value_counts(dropna = False)

# There are some entries with no region name assigned. To address this, making a subset with NaN in the 'region' column
region_NaN_df = df_names_dropped[df_names_dropped['region'].isnull() == True]

region_NaN_df[['state', 'region']].head(100)

# To see if New Hampshire is the only state in this subset dataframe with NaN assigned in 'region' column, using value count.
region_NaN_df['state'].value_counts(dropna = False)

# This confirms that there is a problem with the 'New Hampshire' spelling in the original list named Northeast_states.
# Originally it was misspelled as 'New Hamphire', fixing the spelling now.
Northeast_states = ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 
                    'New York', 'Pennsylvania', 'New Jersey']

# Running loc() function again for Northeast region
df_names_dropped.loc[df_names_dropped['state'].isin(Northeast_states), 'region'] = 'Northeast'

# Checking once again if any missing values in 'region' column
df_names_dropped['region'].value_counts(dropna = False)

# New 'region' column has been successfully created.

To determine whether there's a difference in spending habits between the different regions, creating crosstabs of the variables 'region' and 'spending flag' columns. 

crosstab = pd.crosstab(df_names_dropped['region'], df_names_dropped['spender_type_flag'], dropna = False)

crosstab

# NOTE: The numbers above are NOT YET separated nor grouped by user_id, so how many of each kind of spender is inaccurate since 
# every row is by order_id, which is each product purchased. 

Percent of transactions made by high spenders in each region:

mw_percent = 155975/(7441350+155975)

print(mw_percent)

Midwest: 2.05%

ne_percent = 108225/(108225+5614511)

print(ne_percent)

Northeast: 1.89%

s_percent = 209691/(209691+10582194)

print(s_percent)

South: 1.94%

w_percent = 160354/(160354+8132559)

print(w_percent)

West: 1.93%

The difference of spending habits between the regions is very small. Of the four regions, the Midwest had the highest proportion of transactions made by customers that categorized as 'High spenders.'

We want to only focus on customers who generate regular revenue for the app. Creating an exclusion flag for low-activity customers with less than 5 orders, and excluding them from the dataset. 

# We already have the column 'max_orders' showing the maximum order made by the user. With this, making a exclusion flag. 
df_names_dropped.loc[df_names_dropped['max_order'] < 5, 'exclusion_flag'] = 'Low Activity Customer'

df_names_dropped.loc[df_names_dropped['max_order'] >= 5, 'exclusion_flag'] = 'Regular Activity Customer'

df_names_dropped['exclusion_flag'].value_counts(dropna = False)

# Creating a new dataframe with low-activity customers excluded
df_regular_activity = df_names_dropped[df_names_dropped['exclusion_flag'] == 'Regular Activity Customer']

# Checking to see if new dataframe only includes 'Regular Activity Customer'
df_regular_activity['exclusion_flag'].value_counts(dropna = False)

# Exporting this dataframe
df_regular_activity.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'reg_act_customers.pkl'))
