==> PythonScripts/10.2_FinalAnalysis.ipynb <==
# 10.2 Final Analysis 

#### Import libraries and dataframe - input 1
#### Tidying dataframe: drop columns and update datatype - input 4
#### Age-range profile variables - input 20 
#### Income-range profile variable - input 31
#### Parental-status profile variable - input 48
#### Updating new variables datatypes - input 53
#### Crosstabs
#### Exporting updated dataframe


Importing libraries and dataframe

import pandas as pd
import numpy as np
import os

path = path = r'C:\Users\Neena Tilton\Dropbox\Projects\01_2020_InstacartBasket'

main_df = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'reg_act_customers.pkl'))

Tidying some of the data:

main_df.info()

Columns that need attention: 13, 14, 15, 17, 18

pd.options.display.max_rows = None

main_df[['_merge', 'busiest_day', 'Busiest_days', 'max order', 'max_order']].head(100)

# '_merge' is no longer needed. Columns 'busiest_day' and 'Busiest_days' have different values. Seems like 'max order' 
# and 'max_order' have the same information. 
df_main = main_df.drop(columns = ['_merge', 'max order'])

df_main.info()

# Changing datatype of categorical columns from string to category
df_main['busiest_day'] = df_main['busiest_day'].astype('category')

df_main['Busiest_days'] = df_main['Busiest_days'].astype('category')

df_main['busiest_period_of_day'] = df_main['busiest_period_of_day'].astype('category')

df_main['loyalty_flag'] = df_main['loyalty_flag'].astype('category')

df_main['spender_type_flag'] = df_main['spender_type_flag'].astype('category')

df_main['customer_frequency_flag'] = df_main['customer_frequency_flag'].astype('category')

df_main['gender'] = df_main['gender'].astype('category')

df_main['fam_status'] = df_main['fam_status'].astype('category')

df_main['region'] = df_main['region'].astype('category')

df_main.info()

# Dropping columns decreased memory usage from 7.6+ GB to 7.4+ GB
# Changing string data types to category decreased memory usage from 7.4+ GB to 5.6+ GB. 

Creating new customer profiling variables for further insights.  
Creating age range profile:

df_main['age'].max()

df_main['age'].min()

df_main.loc[df_main['age'] < 20, 'profile_age'] = '19 & under'

df_main.loc[(df_main['age'] >= 20) & (df_main['age'] < 30), 'profile_age'] = '20s'

df_main.loc[(df_main['age'] >= 30) & (df_main['age'] < 40), 'profile_age'] = '30s'

df_main.loc[(df_main['age'] >= 40) & (df_main['age'] < 50), 'profile_age'] = '40s'

df_main.loc[(df_main['age'] >= 50) & (df_main['age'] < 60), 'profile_age'] = '50s'

df_main.loc[(df_main['age'] >= 60) & (df_main['age'] < 70), 'profile_age'] = '60s'

df_main.loc[(df_main['age'] >= 70) & (df_main['age'] < 80), 'profile_age'] = '70s'

df_main.loc[df_main['age'] >= 80, 'profile_age'] = '80 & over'

df_main['profile_age'].value_counts(dropna = False)

Creating income bracket profile:

df_main['income'].min()

df_main['income'].max()

df_main.loc[df_main['income'] < 30000, 'profile_income'] = '25K ~ 29K'

df_main.loc[(df_main['income'] >= 30000) & (df_main['income'] < 40000), 'profile_income'] = '30K ~ 39K'

df_main.loc[(df_main['income'] >= 40000) & (df_main['income'] < 50000), 'profile_income'] = '40K ~ 49K'

df_main.loc[(df_main['income'] >= 50000) & (df_main['income'] < 60000), 'profile_income'] = '50K ~ 59K'

df_main.loc[(df_main['income'] >= 60000) & (df_main['income'] < 70000), 'profile_income'] = '60K ~ 69K'

df_main.loc[(df_main['income'] >= 70000) & (df_main['income'] < 80000), 'profile_income'] = '70K ~ 79K'

df_main.loc[(df_main['income'] >= 80000) & (df_main['income'] < 90000), 'profile_income'] = '80K ~ 89K'

df_main.loc[(df_main['income'] >= 90000) & (df_main['income'] < 100000), 'profile_income'] = '90K ~ 99K'

df_main.loc[(df_main['income'] >= 100000) & (df_main['income'] < 150000), 'profile_income'] = '100K ~ 149K'

df_main.loc[(df_main['income'] >= 150000) & (df_main['income'] < 200000), 'profile_income'] = '150K ~ 199K'

df_main.loc[(df_main['income'] >= 200000) & (df_main['income'] < 300000), 'profile_income'] = '200K ~ 299K'

df_main.loc[(df_main['income'] >= 300000) & (df_main['income'] < 400000), 'profile_income'] = '300K ~ 399K'

df_main.loc[(df_main['income'] >= 400000) & (df_main['income'] < 500000), 'profile_income'] = '400K ~ 499K'

df_main.loc[df_main['income'] >= 500000, 'profile_income'] = '500K ~ 599K'

df_main['profile_income'].value_counts(dropna = False)

Creating parental status profile: 

df_main.loc[df_main['n_dependants'] < 1, 'profile_parental'] = 'NOT Parent'

df_main.loc[df_main['n_dependants'] >= 1, 'profile_parental'] = 'Parent'

df_main.shape

df_main['profile_parental'].value_counts(dropna = False)

23224883 +  7739681

Changing data types of newly created profile columns to category.

df_main['profile_age'] = df_main['profile_age'].astype('category')

df_main['profile_income'] = df_main['profile_income'].astype('category')

df_main['profile_parental'] = df_main['profile_parental'].astype('category')

df_main.info()

Creating crosstabs for more insight:
Crosstab: single-parent status 

xtab_singl_p = pd.crosstab(df_main['fam_status'], df_main['profile_parental'], dropna = False)

xtab_singl_p

xtab_singl_p.to_clipboard()

Crosstab: loyalty & age-range/region/income/parental

# loyalty and age-range
xtab_loyalty_age = pd.crosstab(df_main['loyalty_flag'], df_main['profile_age'], dropna = False)

xtab_loyalty_age

xtab_loyalty_age.to_clipboard()

# loyalty and region
xtab_loyalty_region = pd.crosstab(df_main['loyalty_flag'], df_main['region'], dropna = False)

xtab_loyalty_region

xtab_loyalty_region.to_clipboard()

# loyalty and income-range
xtab_loyalty_income = pd.crosstab(df_main['loyalty_flag'], df_main['profile_income'], dropna = False)

xtab_loyalty_income

xtab_loyalty_income.to_clipboard()

# loyalty and parental-status
xtab_loyalty_parental = pd.crosstab(df_main['loyalty_flag'], df_main['profile_parental'], dropna = False)

xtab_loyalty_parental

xtab_loyalty_parental.to_clipboard()

Crosstab: customer-frequency and age/gender/parental/income/fam-status

# customer-frequency and age-range
xtab_cstfrqncy_age = pd.crosstab(df_main['customer_frequency_flag'], df_main['profile_age'], dropna = False)

xtab_cstfrqncy_age

xtab_cstfrqncy_age.to_clipboard()

# customer-frequency and gender
xtab_cstfrqncy_gender = pd.crosstab(df_main['customer_frequency_flag'], df_main['gender'], dropna = False)

xtab_cstfrqncy_gender

xtab_cstfrqncy_gender.to_clipboard()

# customer-frequency and parental-status
xtab_cstfrqncy_parent = pd.crosstab(df_main['customer_frequency_flag'], df_main['profile_parental'], dropna = False)

xtab_cstfrqncy_parent 

xtab_cstfrqncy_parent.to_clipboard()

# customer-frequency and income-range
xtab_cstfrqncy_income = pd.crosstab(df_main['customer_frequency_flag'], df_main['profile_income'], dropna = False)

xtab_cstfrqncy_income 

xtab_cstfrqncy_income.to_clipboard()

# customer-frequency and family-status
xtab_cstfrqncy_famstatus = pd.crosstab(df_main['customer_frequency_flag'], df_main['fam_status'], dropna = False)

xtab_cstfrqncy_famstatus

xtab_cstfrqncy_famstatus.to_clipboard()

Crosstab: spending-type and age/gender/income/fam-status/parental-status 

# spender-type and age
xtab_spndrtype_age = pd.crosstab(df_main['spender_type_flag'], df_main['profile_age'], dropna = False)

xtab_spndrtype_age

xtab_spndrtype_age.to_clipboard()

# spender-type and gender
xtab_spndrtype_gender = pd.crosstab(df_main['spender_type_flag'], df_main['gender'], dropna = False)

xtab_spndrtype_gender 

xtab_spndrtype_gender.to_clipboard() 

# spender-type and income
xtab_spndrtype_income = pd.crosstab(df_main['spender_type_flag'], df_main['profile_income'], dropna = False)

xtab_spndrtype_income

xtab_spndrtype_income.to_clipboard() 

# spender-type and family-status
xtab_spndrtype_famstatus = pd.crosstab(df_main['spender_type_flag'], df_main['fam_status'], dropna = False)

xtab_spndrtype_famstatus

xtab_spndrtype_famstatus.to_clipboard() 

# spender-type and parental-status
xtab_spndrtype_parental = pd.crosstab(df_main['spender_type_flag'], df_main['profile_parental'], dropna = False)

xtab_spndrtype_parental

xtab_spndrtype_parental.to_clipboard() 

Crosstab: order-dow and age/parental-status/fam-status

# order day-of-week and age
xtab_dow_age = pd.crosstab(df_main['orders_day_of_week'], df_main['profile_age'])

xtab_dow_age

xtab_dow_age.to_clipboard()

# order day-of-week and parental-status
xtab_dow_parental = pd.crosstab(df_main['orders_day_of_week'], df_main['profile_parental'])

xtab_dow_parental

xtab_dow_parental.to_clipboard()

# order day-of-week and family-status
xtab_dow_famstatus = pd.crosstab(df_main['orders_day_of_week'], df_main['fam_status'])

xtab_dow_famstatus 

xtab_dow_famstatus.to_clipboard()

Crosstab: product_id and age/income/fam-status/parent-status/gender

# product-id and age-range
xtab_prodID_age = pd.crosstab(df_main['product_id'], df_main['profile_age'])

xtab_prodID_age

xtab_prodID_age.to_clipboard()

# product-id and income-range
xtab_prodID_income = pd.crosstab(df_main['product_id'], df_main['profile_income'])

xtab_prodID_income.to_clipboard()

# product-id and family-status
xtab_prodID_famstatus = pd.crosstab(df_main['product_id'], df_main['fam_status'])

xtab_prodID_famstatus.to_clipboard()

# product-id and parental-status
xtab_prodID_parental = pd.crosstab(df_main['product_id'], df_main['profile_parental'])

xtab_prodID_parental.to_clipboard()

# product-id and gender
xtab_prodID_gender = pd.crosstab(df_main['product_id'], df_main['gender'])

xtab_prodID_gender.to_clipboard()

Crosstab: department_id and age/income/fam-status/parent-status/gender

# department_id and age
xtab_dptmt_age = pd.crosstab(df_main['department_id'], df_main['profile_age'], dropna = False)

xtab_dptmt_age

xtab_dptmt_age.to_clipboard()

# department_id and income-range
xtab_dptmt_income = pd.crosstab(df_main['department_id'], df_main['profile_income'], dropna = False)

xtab_dptmt_income

xtab_dptmt_income.to_clipboard()

# department_id and family-status
xtab_dptmt_famstatus = pd.crosstab(df_main['department_id'], df_main['fam_status'], dropna = False)

xtab_dptmt_famstatus

xtab_dptmt_famstatus.to_clipboard()

# department_id and parental-status
xtab_dptmt_parental = pd.crosstab(df_main['department_id'], df_main['profile_parental'], dropna = False)

xtab_dptmt_parental

xtab_dptmt_parental.to_clipboard()

# department_id and gender
xtab_dptmt_gender = pd.crosstab(df_main['department_id'], df_main['gender'], dropna = False)

xtab_dptmt_gender

xtab_dptmt_gender.to_clipboard()

Crosstab: age and income

xtab_age_income = pd.crosstab(df_main['profile_age'], df_main['profile_income'], dropna = False)

xtab_age_income

xtab_age_income.to_clipboard()

Crosstab: region and income

xtab_region_income = pd.crosstab(df_main['region'], df_main['profile_income'], dropna = False)

xtab_region_income

xtab_region_income.to_clipboard()

Exporting main dataframe

df_main.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'df_main.pkl'))
